Rework actions update
