+++ apps/openmw/mwinput/keyboardmanager.cpp
@@ -27,70 +27,9 @@ namespace MWInput
     {
     }
 
-    bool KeyboardManager::update(float dt, bool disableControls)
+    void KeyboardManager::update(float dt, bool disableControls)
     {
         mControlsDisabled = disableControls;
-
-        // Disable movement in Gui mode
-        if (MWBase::Environment::get().getWindowManager()->isGuiMode()
-            || MWBase::Environment::get().getStateManager()->getState() != MWBase::StateManager::State_Running)
-        {
-            return false;
-        }
-
-        // Configure player movement according to keyboard input. Actual movement will
-        // be done in the physics system.
-        if (!MWBase::Environment::get().getInputManager()->getControlSwitch("playercontrols"))
-        {
-            return false;
-        }
-
-        bool triedToMove = false;
-        bool alwaysRunAllowed = false;
-
-        MWWorld::Player& player = MWBase::Environment::get().getWorld()->getPlayer();
-
-        if (actionIsActive(A_MoveLeft) != actionIsActive(A_MoveRight))
-        {
-            alwaysRunAllowed = true;
-            triedToMove = true;
-            player.setLeftRight (actionIsActive(A_MoveRight) ? 1 : -1);
-        }
-
-        if (actionIsActive(A_MoveForward) != actionIsActive(A_MoveBackward))
-        {
-            alwaysRunAllowed = true;
-            triedToMove = true;
-            player.setAutoMove (false);
-            player.setForwardBackward (actionIsActive(A_MoveForward) ? 1 : -1);
-        }
-
-        if (player.getAutoMove())
-        {
-            alwaysRunAllowed = true;
-            triedToMove = true;
-            player.setForwardBackward (1);
-        }
-
-        if (triedToMove)
-            MWBase::Environment::get().getInputManager()->resetIdleTime();
-
-        static const bool isToggleSneak = Settings::Manager::getBool("toggle sneak", "Input");
-        if (!isToggleSneak)
-        {
-            if(!MWBase::Environment::get().getInputManager()->joystickLastUsed())
-                player.setSneak(actionIsActive(A_Sneak));
-        }
-
-        float xAxis = mInputBinder->getChannel(A_MoveLeftRight)->getValue();
-        float yAxis = mInputBinder->getChannel(A_MoveForwardBackward)->getValue();
-        bool isRunning = xAxis > .75 || xAxis < .25 || yAxis > .75 || yAxis < .25;
-        if ((mActionManager->isAlwaysRunActive() && alwaysRunAllowed) || isRunning)
-            player.setRunState(!actionIsActive(A_Run));
-        else
-            player.setRunState(actionIsActive(A_Run));
-
-        return triedToMove;
     }
 
     bool KeyboardManager::actionIsActive (int id)
